requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.118
botocore==1.29.118
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.6
datasets==2.11.0
dill==0.3.6
docker==6.0.1
entrypoints==0.4
filelock==3.12.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mlflow==2.3.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyarrow==11.0.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
regex==2023.3.23
requests==2.28.2
responses==0.18.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.10
sqlparse==0.4.4
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.9.1
zipp==3.15.0
